Introduz navegação, para o mapa ficar numa view separada
