+++ package.json
@@ -24,6 +24,7 @@
     "expo": "^21.0.0",
     "react": "16.0.0-alpha.12",
     "react-native": "^0.48.4",
-    "react-native-dotenv": "^0.1.0"
+    "react-native-dotenv": "^0.1.0",
+    "react-navigation": "^1.0.0-beta.15"
   }
 }
