+++ src/screens/Line.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { MapView } from 'expo';
+
+export default class Line extends React.Component {
+  render () {
+    return (
+      <MapView
+        initialRegion={{
+          latitude: 37.78825,
+          longitude: -122.4324,
+          latitudeDelta: 0.0922,
+          longitudeDelta: 0.0421,
+        }}
+      />
+    );
+  }
+}
